// src/app/modules/configuration/general/models/class-result-view-model.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * ClassViewModel
 */
export interface ClassResultViewModel {
    /**
     * Gets or Sets Id
     */
    id: string;
    /**
     * Gets or Sets ClassReferenceName
     */
    classReferenceName?: string;
    /**
     * Gets or Sets Name
     */
    name: string;
    /**
     * Gets or Sets ClassReferenceTypeId
     */
    classReferenceTypeId: string;
    /**
     * Gets or sets Code
     */
    code?: string;
    /**
     * Gets or sets Description
     */
    description?: string;
}

/**ValidationMessageView */
export interface ValidationMessageView {
    /**///  */
    // tslint:disable-next-line:no-misused-new
    new(): ValidationMessageView;
    /**Gets or sets list of string mesaages */
    messages: { [key: string]: string };

    /**Gets or sets StatusCode */
    statusCode: number;
}

export interface PagedList<T> {
    // public PagedList(IQueryable<T> source,pageNumber: number;

    totalItems: number;
    pageNumber: number;
    pageSize: number;
    list: T[];
    totalPages: number;
    hasPreviousPage: boolean;
    hasNextPage: boolean;
    nextPageNumber: number;
    previousPageNumber: number;

    // public PagingHeader GetHeader();
}

/**ClassResultView to return List of Classes with Validation */
export interface ClassResultView extends ValidationMessageView {
    /**Gets or Sets Paged List of ClassViewModels */
    pagedClassViewModels: PagedList<ClassResultViewModel>;
    filters: Array<any>;
}